Move some live_repair tests to downstairs.rs
